+++ utils/CellRefUtil.js
@@ -93,9 +93,9 @@ var CellRefUtil;
     function mergeCellSpans(offsetIdx, length, spansStr) {
         var min1 = offsetIdx + 1;
         var max1 = offsetIdx + length;
-        var _a = spansStr.split(":"), min2 = _a[0], max2 = _a[1];
-        min2 = parseInt(min2);
-        max2 = parseInt(max2);
+        var _a = spansStr.split(":"), min2Str = _a[0], max2Str = _a[1];
+        var min2 = parseInt(min2);
+        var max2 = parseInt(max2);
         var min = Math.min(min1, min2);
         var max = Math.max(max1, max2);
         return { spans: min + ":" + max, min: min, max: max }; // 1 based, inclusive